feat: build training data (#7)
At this stage, features and targets can be created with FeatureExtractor and TargetProcessor modules. This PR aims at creating the first dataset and training the first model.

The process is the following:

Extract features and targets using FeatureExtractor and TargetProcessor.
Upload them as pickle files to gDrive.
From there, the data can be loaded in Google Colab for training.